FInished
Finished the whole project.

+++ pages/api/orders/[id].js
@@ -2,9 +2,12 @@ import dbConnect from '../../../util/mongo'
 import Order from '../../../models/Order'
 
 const handler = async (req,res) =>{
-    const { method, query:{id} } = req;
+    const { method, query:{id}, cookies } = req;
 
     dbConnect()
+
+    const myCookie = cookies?.token || ''
+    const admin = myCookie===process.env.TOKEN?true:false
 
     switch(method){
         case "GET":
@@ -17,6 +20,7 @@ const handler = async (req,res) =>{
             }
             break;
         case "PUT":
+            if (!admin) return res.status(401).json("Not authenticated")
             try{
                 const order = await Order.findByIdAndUpdate(id,req.body,{
                     new:true, // return newest version
@@ -28,6 +32,7 @@ const handler = async (req,res) =>{
             }
             break;
         case "DELETE":
+            if (!admin) return res.status(401).json("Not authenticated")
             try{
                 const order = await Order.findByIdAndDelete(id)
                 res.status(200).json(`Order with id ${id} has been deleted successfully!`)
